+++ src/WebApiBolierplate/Core.Constants/InfoTexts.cs
@@ -1,6 +1,0 @@
-﻿namespace Core.Constants;
-
-public static class InfoTexts
-{
-    public const string Success = "Success";
-}