+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Removed
  - Base class for operations is removed
+ - Removed unused constants
 
 ## [7.0.0] - 2023-02-27
 ### Added